detect: Saving wwn & scsi IDs separately
When getting disk's properties, we were saving the scsi & wwn IDs in the
same variable name called "id".

That wasn't really useful as in bug like #139, we should be in position
to identify the wwn of a particular disk to avoid disk's renaming.

Using the wwn IDs allow you to keep a constant path to reach this
particular disk by using the /dev/disk/by-id/<wwn-id> path.

Thanks to this patch, you can now write such matching rule in a
spec:

....
('disk', '$first_disk', 'www-id', '$fist_disk_wwn_id')
....

In a configure script, instead of doing the usual:
  my_disk_path = '/dev/ + var['first_disk']

You can write
  my_disk_path = '/dev/disk/by-id/' + var['first_disk_wwn_id']

This is perfectly targeting the same disk, the 2nd way of writing the
code makes it more robust to dynamic disks renaming like reported in the
issue #139

This fix is closing #139