init: Adding LLDP support to init
This patch is one the serie to get LLDP information being reported by eDeploy.

This patch is making the lldp service running and waits until the lldp
status is reported as ok.

It will wait up to 45 seconds to get this LLDP report being done.

As the LLDP is started at the same time as the nic configuration, we
have a good chance the NIC will have received some LLDP packets during
the DHCP requests.