detect: Adding physical_block_size & optimal_io_size for disks
When detecting disks, it could be very useful to get the optimal_io_size
and the physical_block_size of the disk device to compute a good
alignement.

Some tools or setups, like for setting a raid array, requires this
information as an input (stripe-size for example).

By adding those two information, we let the ability to reuse them into
the configure.

A typical output looks like:
 ('disk', 'sda', 'optimal_io_size', '0'),
 ('disk', 'sda', 'physical_block_size', '512'),