detect: Getting hpa disk's size from pd show
Up to now, physical's disk size of an hpa controller was taken by
reading the "pd all" command leading to problems like in issue #134

As we do have the information from the "pd show" command, it's safer and
easier to gain the information from it.

This patch is now reading the pd size from the pd_show instead of
pd_all.

This is a fix for issue #134