health: Adding required-hypervisors support
While benchmarking VMs on an infrastructure, it is important to insure
they run on the expected number of hypervisors.

Using 'required-hypervisors' feature enforce a check and prevent the
benchmark to start if too few hypervisors got detected.

Networking tests are very sensible to this parameter as of the arity
setting.

This parameter is really important to avoid benchmarking the proper
amount of VMs on a reduced set of hypervisors.