init: Improving error message if no DHCP answer
When an interface was up and failed to get a DHCP answer the message
was: "$iface configuration failed"

When reading the log message it was pretty confusing because you don't
really know if the setup of the interface failed or the dhcp answer was
missing.

As this message is triggered by the latter case, let's print it this way
: "No DHCP answer received on $iface"

This way we really know that the interface setting was correct but just
missed of a DHCP answer.