init: Don't report too early failure
A real system reported the following log :
 Waiting for 1 DHCP anwsers to come up in 44 sec
 No DHCP answer received on eth1
 Found 1 interfaces properly configured

This is pretty confusing because we did receive an answer but said it's
not the case.

We were trying to detect the same thing twice by two different methods
leading to inconsistent error messages.

Let's keep only the second print which was right. As we also do print
the "ip a" command, that could let user reading what was really
configured.