repositories: Attach pool if not already done + detach function
Currently whenever a call is made to attach_pool_rh_cdn(), edeploy
tries to attach to it even though it has already been attached, leading
to a fail build. This commit aims to verify if the pool is already attached
before attaching to it. This commit also provide a function detach_poll_rh_cdn()