detect: Adding AHCI flags detection
AHCI features could be very different regarding the bios setup.
Let's detect the flags from the dmesg to report them. This way we can
detect such difference at cardiff time.

This commit fixes issue #150.

Note that we store the dmesg in /tmp/dmesg for performance issues.
We'll surely have to add more dmesg parsing, so reading a file on the
filesystem will be faster than calling dmesg all the time.