cardiff: Fixing trace names while using multiple rampup dirs
When using the rampup option, you can provide many tests as input.

In such case, the existing code was mixing data and trace name. As a
result, we could have inverted results like plotting a low performance
result with the title of the one that was good. This is very misleading.

This fix insure that traces name are taken in the same order as the data
are plot into the trace files.