detect: Split Inquiry data in subfields
The Inquery data is very complicated to diff because it usually contains
a string composed of three parts : vendor, product, serial.

___but____ the order is pretty undefined and seems to be vendor
specific.

To ease the product/vendor comparaison, we now split the "Inquiry Data"
in sub fields. It will be much more easier to compare sub fields instead
of one big string.

Prior this patch, we had an output like :
 ('pdisk', 'disk1', 'InquiryData[0]', 'SEAGATE    ST300MM0006   LS08S0K1Z3TM'),

Now a typical output looks like:
 ('pdisk', 'disk1', 'InquiryData[0]', 'SEAGATE'),
 ('pdisk', 'disk1', 'InquiryData[1]', 'ST300MM0006'),
 ('pdisk', 'disk1', 'InquiryData[2]', 'LS08S0K1Z3TM'),

It will be up to humans to consider which fields are important to track.