[upgrade] exclude /usr/share/ansible on install-server
From I.1.2.X to I.3.0 releases, Ansible has been upgraded and because of
this commit:
https://github.com/ansible/ansible/commit/ae2d198342c9bbb691647b897b69b9d9f26522db

The playbook that is run on the tag install_server will stop & fail
after eDeploy upgrade task because it won't find the necessary Ansible
library code that has been moved.

To avoid that, let's exclude this directory for this version but later
this directory may be dropped from the exclude file.