build/base: also fake useradd/groupadd after installing shadow-utils
Before that patch, we used to fake useradd/groupadd before installing
base packages.

"shadow-utils" is part of these basic packages so useradd/groupadd were
overwritten and we could not fake the commands correctly, so all the
UID/GID were not set accordingly with /root/ids.tables so the upgrades
could not work correctly.

We did not have the issue on Debian because they are already installed
when we want to fake the commands (within debootstrap).

This patch transforms the faking in a function that is used two times to
ensure that we track a maximum of users/groups creation between
bootstrap and after that shadow-utils is installed.
After shadow-utils installed, we don't have to worry anymore since the
fake commands will never move.