build/check-ug.py: UID/GID Validator
This script is intended to be run at the end
of an image creation by eDeploy. This is still
a WIP but can be merge like that as it does not
introduce modifications in the image creation
process.

It verify that all users and groups that all rpm
requested are really created and with the expected
ids/gids according to ids.tables.

This is an helper to avoid unexpected behaviors
due to eventual problems in mngids.py wrapper that are
silently skipped by pre-inst scripts of rpms.