init: Avoid special char in filename
While converting the DMI information into variable, we shall be sure
that no special char will be present.

Priori, this commit we had a case where the system-product-name was : "Standard PC (i440FX + PIIX, 1996)",
as a result we had a filename containing a coma leading to an error
reported in issue #203.

This patch does remove all special char and turn them into underscore
(_).

That should be enough to kick this issues.

The current code generates : "Standard_PC_i440FX_PIIX_1996" for the current case.