init: Reporting if PXEMGNR was succesfull or not
We had a case were the PXEMGNR didn't work on the server side by
returning a 404. We had no proof during the deploiement that it fails as
no message was printed in the log file.

This patch does make an explicit report if it succeed or failed to allow
such analysis.