init: Give up if unable to get a lock on edeploy server
As reported in issue #169, if some error occurs on the server side,
client is connected but cannot get a configure file and stay in this
waiting loop forever.

This is hard to debug/understand because user doesn't have any shell to
understand this situation. Even worse, that could be understood as a
client issue while in fact it's a server issue.

This commit is so implementing a default timeout of 2 minutes before
giving up, print an explicit message and give a shell to the user.