upload.py: Avoid process waiting for user input
As per issue #202, we could be in a situation where a process is waiting
for a particular user input.

That's weird because user cannot see the message, install process is
locked and doesn't provide a shell because the command didn't failed at
that time.

The idea proposed by @fredericlepied is to remove the stdin making
applications waiting for a stdin to fail.

I tried on a prototype and that worked perfectly.