doc: Adding default values for boot parameters
eDeploy have plenty boot parameters that can be passed on the command
line. They are all documented but as the default value isn't reported,
it's pretty hard to get what is the default behavior of eDeploy.

This patch does add a column showing the default value for each
parameter making each entry more explicit.