health: Don't override serial value
The serial value is defined by get_system_info() in init.common but
init.health still had a line that overloaded that value.

As a result, the serial variable didn't had all the protections against
white spaces, special char, ... That was preventing in those cases to
upload the log file to the server.

This commit just remove the useless override in init.health, making upload_log() now
using the proper $serial variable.