deploy: Provide additional localboot methods
As per issue #224, we have servers, more likely HP ones, that like not
behaving well when pxelinux ended his works and trigger localboot.

Regarding the crack smokers bios developpers, you have to search which
one of the following workaround to make a localboot :
- exiting properly aka localboot 0
- exiting in failure aka localboot -1
- chainloading directly the disk by using chain.c32

Up to now, we were only providing one option in the default
configuration file.

This patch does:
- add the missing chain.c32 file
- add potential other localboot workaround

It will be up to users testing them and select the appropriate one. At
least, we give a serious hint to make this task being fast.

Thanks @AlexandreNo for reporting that.