base: Cleaning InstallMedia RPM cache
As reported in issue #132 of edeploy-roles, the base role let many rpms
in the cache directory of yum even if the "yum clean all" got called.

All those rpms are part of the initial "base" install during the
bootstrapping and at that time the yum didn't track those rpms and so
isn't able to clean them.

This patch is kind of a workaround that enforce the deletion of all
remaining rpms located in the yum cache.

That's not perfect but that will save a lot of disk space.