RPM based-images: generate pre/post metadata scripts
This patch aims to automatically generate pre/post scripts so at each
upgrade we can run these scripts to complete the rsync process and
ensure our images from a fresh install.
A typical use-case is selinux which uses to execute new context at each
upgrade, so with this patch we won't miss this upgrade anymore and
automatically execute it durng `edeploy upgrade` process.

* generate eDeploy metadatas pre/post scripts.
  During the build, generate pre/post scripts and copy them in the image
  (in /var/lib/edeploy/<version> directory).
  Note: the script only supports RPM based systems.
* drop metadata/RH7.0-J.1.0.0/*/RH7.0-J.1.1.0/{pre,post} (not useful
  anymore, since they are generated)

Co-Authored-By: Frederic Lepied <frederic.lepied@enovance.com>