init: Adding grub_options() helper
When a particular role needs to setup a kernel option, it was pretty tricky to
do it without colliding with all the config done by "init" script.

This patch adds an helper that can be called directly from the configure script
to add a particular boot option.

A typical usage looks like :

    grub_options('net.ifnames=0 rw')

This options are installed in the grub configuration but are also used whenever
the kexec boot option is used.