pxe: Adding package dependencies
The pxe targets uses `hardware` project which have pandas as build dependency.
Python is part of it and takes ages for being installed by pip as it recompiles
a lot of code in a single threaded way.

This patch adds then python-pandas as default dep and adds also
python-setuptools for debian-based distro as it was missing.