init: Properly detect kexec's root uuid when grub2 is used
When trying to kexec, we need to have the proper root uuid but the current code
was only considering the grub1 case.

This patch add the detection of the uuid when using grub2.

As the grub version is already detected and stored into the $V variable, let's
reuse it here.