init: Fixing lldpad respawning
When using the respawn command, the lldp tooling failed generating false
negative matches.

The main reason is that /run was umounted by the "umount -a" but never been
remounted as /dev/ is kept during the respawn.

This patch moves the tmpfs mount to be done inconditionaly but also kill lldap
properly.