health-check: Avoid destructing mounted partitions
When benchmarking disks in DESTRUCTIVE_MODE, it's mandatory not to test mounted filesystems.
As the benchmark can be run on a alive server, is_booted_storage_device() isn't sufficient.

The new proposal is to detect if a give block device have any mounted filesystem on it.
This is done by using the '--fs' option of lsblk.

is_booted_storage_device() is so renamed to is_mounted_storage_device()

Signed-off-by: Erwan Velu <e.velu@criteo.com>