health-check: Adding numa pinned benchmarks
When performing benchmarks on numa aware systems, the memory bandwidth could be dramatically improved when pinning processes to numa node.
This patch is about to run the memory benchmark but per numa node by loading the appropriate number of cores in it.
The numa pinning and cpu masks are extracted from the python-hardware library.
health-check will populate <n> threads (where n is the number of numa nodes) and will spawn 1 instance of sysbench with <m> threads (where m is the number of cpu cores per numa node).
The affinity between the sysbench instance and the numa node is done by using the proper cpu mask when calling taskset.

On massively numa-aware architectures like AMD EPYC 7451, the performance improve is really important.

Signed-off-by: Erwan Velu <e.velu@criteo.com>